feat: 단위 변경

+++ app/src/main/java/com/example/weathercard/APIData/Atmosphere.java
@@ -1,20 +1,20 @@
 package com.example.weathercard.APIData;
 
 public class Atmosphere {
-    private long humidity;
-    private long visibility;
+    private double humidity;
+    private double visibility;
     private double pressure;
-    private long rising;
+    private double rising;
 
-    public long getHumidity() { return humidity; }
+    public double getHumidity() { return humidity; }
     public void setHumidity(long value) { this.humidity = value; }
 
-    public long getVisibility() { return visibility; }
+    public double getVisibility() { return visibility; }
     public void setVisibility(long value) { this.visibility = value; }
 
     public double getPressure() { return pressure; }
     public void setPressure(double value) { this.pressure = value; }
 
-    public long getRising() { return rising; }
+    public double getRising() { return rising; }
     public void setRising(long value) { this.rising = value; }
 }
